- updated MapsLTProjection, they switch back to 256x25 tiles again ;} - fixed waitOnEmptyTasks.WaitOne, GMap.NET.WindowsForms now should run without SP2 in .NET2 - fixed zoom bug in SetZoomToFitRect

+++ GMap.NET.WindowsForms/System.Windows.Forms/GMapControl.cs
@@ -382,7 +382,7 @@ namespace System.Windows.Forms
                maxZoom = MaxZoom;
             }
 
-            if(Zoom != maxZoom)
+            if((int)Zoom != maxZoom)
             {
                Zoom = maxZoom;
             }
